move regional catalog helpers to RegionalCatalogEntry
